redesigned JavaScript, AJAX response with post_id
I would like to use WP-like-System on pages which listed more than one articles or pages.
The Scripts is now listening on click events on all elements with the class ".btn-like".
If on of those elements is clicked, a AJAX request is send to the server which responses in addition with the postId of the liked item.
The script than changes only the "btn-like"-element with the matching postId.